requirements.txt
google-cloud-aiplatform
PyPDF2
asyncio  # Built-in, but for clarity
tenacity